Fixed (lack of) error handling on `next` URLs
To avoid duplicate code, I combined the requesting + error handling of the original API call, and, if applicable, any subsequent `next` API calls. However, the consequence of this is that the code probably makes no sense to the untrained eye. Hence, the absolute barrage of comments.
(I promise you, the tests *do* pass!)